fix: fingerprints and edge cases
fix: future-proof src organization
