+++ sdcli/blackstrap/README.md
@@ -1,0 +1,5 @@
+# blackstrap
+
+Blackstrap is an umbrella suite of [skunwork](https://en.wikipedia.org/wiki/Skunkworks_project) projects, tools and services.
+
+Many of them (though not by principal) have CLIs exposed here via `sdcli`.
