+++ sdcli/blackstrap/__init__.py
@@ -1,1 +1,3 @@
-from .s3.blackstrap_s3 import bridge as blackstrap_s3  # noqa
+from .s3.s3 import s3
+
+__all__ = ["s3"]
