--- sdcli/blackstrap/s3/__init__.py
@@ -1,1 +1,0 @@
-from .blackstrap_s3 import bridge  # noqa
