+++ sdcli/commands/__init__.py
@@ -1,0 +1,3 @@
+from .gh import gh
+
+__all__ = ["gh"]
