+++ sdcli/utils.py
@@ -1,8 +1,9 @@
+import hashlib
 import os
 from contextlib import contextmanager
 from pathlib import Path
 from subprocess import PIPE, CalledProcessError, run
-from typing import Optional, Tuple, Union
+from typing import List, Optional, Tuple, Union
 
 import typer
 from cachecontrol import CacheControl
@@ -52,17 +53,17 @@ def wrap_ghsession():
             yield session
     except Exception:
         typer.secho(
-            "\n[ X ] Something went wrong. Here is the traceback:\n",
+            "\n[ X ] Something went wrong communicating with GitHub.\n",
             fg=typer.colors.BRIGHT_RED,
         )
         raise typer.Exit(code=1)
 
 
-def run_command(command: Union[str, list]):
+def run_command(command: Union[str, List[str]], capture: bool = False):
     """
-    Run an arbitrary command with arbitrary arguments. STDOUT is preserved
-    while STDERR is formatted upon unsuccessful command execution, either at
-    the command or OS level.
+    Run an arbitrary command with arbitrary arguments and return the CompletedProcess.
+    STDERR is captured and formatted upon unsuccessful command execution, either at the
+    command or OS level. If capture is True, STDOUT is captured.
     """
     if isinstance(command, str):
         command = command.split(" ")
@@ -70,7 +71,13 @@ def run_command(command: Union[str, list]):
     try:
         # try to the provided command as a subprocess, capturing
         # the stderr if the command fails
-        run(command, universal_newlines=True, stderr=PIPE, check=True)
+        return run(
+            command,
+            text=True,
+            stdout=(capture and PIPE) or None,
+            stderr=PIPE,
+            check=True,
+        )
     except Exception as err:
         # if the command failed, we only care about its stderr
         if isinstance(err, CalledProcessError):
@@ -83,8 +90,8 @@ def run_command(command: Union[str, list]):
         typer.echo(
             typer.style(
                 "\n[ X ] Something went wrong! If you're not a developer,"
-                f" ignore the rest. Otherwise, the traceback from `{' '.join(command)}`"
-                " was recaptured and is printed.",
+                " ignore the rest.\n      Otherwise, the traceback from"
+                f" `{' '.join(command)}` was recaptured and is printed.",
                 fg=typer.colors.BRIGHT_RED,
             )
             + errmsg
@@ -95,15 +102,15 @@ def run_command(command: Union[str, list]):
 def is_docker_supported():
     """Checks if Docker and Docker Compose 2 exist on the system and are running."""
     try:
-        run(["docker", "version"], capture_output=True, check=True)
-        docker_check = run(["docker", "compose", "version"], capture_output=True)
+        run_command("docker version", capture=True)
+        docker_check = run_command("docker compose version", capture=True)
     except (CalledProcessError, FileNotFoundError):
         docker_check = None
 
     if not docker_check or docker_check.returncode != 0:
         typer.secho(
-            "[ X ] Docker Compose is not available but is required. Ensure Docker and"
-            " Docker Compose 2 are installed and running before continuing.",
+            "[ X ] Docker Compose V2 is not available but is required. Ensure Docker is"
+            " running and Docker Compose V2 is installed before continuing.",
             fg=typer.colors.BRIGHT_RED,
         )
         raise typer.Exit(code=1)
@@ -112,26 +119,45 @@ def is_docker_supported():
 def fingerprint_path(
     *service: str,
     fingerprint: Optional[str] = None,
-    hashable: Optional[Tuple[str]] = None,
-) -> Path:
+    hashable: Tuple[Optional[str]] = (),
+) -> Tuple[str, Path]:
     """
     Returns the cache path for a given fingerprint or hashable under the provided
     service.
     """
-    if (not fingerprint and not hashable) or (fingerprint and hashable):
+    any_hashable = any(hashable)
+    if (not fingerprint and not any_hashable) or (fingerprint and any_hashable):
         typer.secho(
             "[ X ] You must supply either the fingerprint of an already configured"
-            " service or a unique identifier for a new fingerprint.",
+            " service or a complete unique identifier for a new fingerprint.",
+            fg=typer.colors.BRIGHT_RED,
+        )
+        raise typer.Exit(code=1)
+    elif any_hashable and not all(hashable):
+        typer.secho(
+            "[ X ] It doesn't seem like you've provided all the arguments required to"
+            " produce a unique fingerprint. Check your specific command usage.",
             fg=typer.colors.BRIGHT_RED,
         )
         raise typer.Exit(code=1)
 
-    path = Path.home().joinpath(".sdcli", *service, fingerprint or str(hash(hashable)))
+    det_fingerprint = (
+        fingerprint
+        # we need predictable results between interpreters, which hash() won't provide
+        or hashlib.md5("|".join(hashable).encode(), usedforsecurity=False).hexdigest()
+    )
+
+    path = Path.home().joinpath(
+        ".sdcli",
+        *service,
+        det_fingerprint,
+    )
 
     if fingerprint and not path.exists():
         typer.secho(
-            "[ X ] The fingerprint provided does not exist.", fg=typer.colors.BRIGHT_RED
+            f"[ X ] The provided fingerprint '{det_fingerprint}' does not exist.",
+            fg=typer.colors.BRIGHT_RED,
         )
         raise typer.Exit(code=1)
 
-    return path
+    return det_fingerprint, path
